[REFACTOR] Move appearance actions to new folder in game logic
