+++ pandora-common/src/assets/modules/lockSlot.ts
@@ -1,16 +1,16 @@
 import { Immutable } from 'immer';
 import { z } from 'zod';
 import { ItemInteractionType } from '../../character/restrictionTypes';
+import type { AppearanceModuleActionContext } from '../../gameLogic/actionLogic/appearanceActions';
 import type { InteractionId } from '../../gameLogic/interactions';
 import { AssertNever, Satisfies } from '../../utility/misc';
-import type { AppearanceModuleActionContext } from '../appearanceActions';
 import { AppearanceItems, AppearanceValidationResult } from '../appearanceValidation';
 import type { Asset } from '../asset';
 import type { AssetManager } from '../assetManager';
 import { ConditionOperator } from '../graphics';
+import { __internal_ItemBundleSchemaRecursive, __internal_ItemTemplateSchemaRecursive } from '../item/_internalRecursion';
 import { IItemCreationContext, IItemLoadContext, IItemValidationContext } from '../item/base';
 import { ItemLock, ItemLockActionSchema } from '../item/lock';
-import { __internal_ItemBundleSchemaRecursive, __internal_ItemTemplateSchemaRecursive } from '../item/_internalRecursion';
 import type { IAssetModuleDefinition, IExportOptions, IItemModule, IModuleActionCommon, IModuleConfigCommon, IModuleItemDataCommon } from './common';
 
 // Fix for pnpm resolution weirdness
