+++ pandora-common/src/assets/modules/typed.ts
@@ -2,9 +2,9 @@ import { Immutable } from 'immer';
 import { z } from 'zod';
 import { CharacterIdSchema } from '../../character/characterTypes';
 import { ItemInteractionType } from '../../character/restrictionTypes';
+import type { AppearanceModuleActionContext } from '../../gameLogic/actionLogic/appearanceActions';
 import type { InteractionId } from '../../gameLogic/interactions';
 import { Satisfies } from '../../utility/misc';
-import type { AppearanceModuleActionContext } from '../appearanceActions';
 import type { AppearanceItems, AppearanceValidationResult } from '../appearanceValidation';
 import type { AssetManager } from '../assetManager';
 import type { ConditionOperator } from '../graphics';
