+++ pandora-common/src/gameLogic/actionLogic/index.ts
@@ -1,0 +1,3 @@
+export * from './appearanceActionProblems';
+export * from './appearanceActionProcessingContext';
+export * from './appearanceActions';
