+++ pandora-common/src/networking/client_shard.ts
@@ -1,20 +1,20 @@
-import { CharacterIdSchema } from '../character/characterTypes';
-import { CharacterPublicSettingsSchema, CharacterRoomPositionSchema } from '../character/characterData';
-import { AppearanceActionSchema } from '../assets/appearanceActions';
-import { AppearanceActionProblem, AppearanceActionData } from '../assets/appearanceActionProblems';
-import { ClientChatMessagesSchema, ChatCharacterStatusSchema } from '../chat/chat';
-import { z } from 'zod';
-import { CharacterInputNameSchema, ZodCast } from '../validation';
-import { Satisfies } from '../utility/misc';
-import { SocketInterfaceDefinition, SocketInterfaceDefinitionVerified, SocketInterfaceHandlerPromiseResult, SocketInterfaceHandlerResult, SocketInterfaceRequest, SocketInterfaceResponse } from './helpers';
 import { Immutable } from 'immer';
-import { PermissionConfigChangeSchema, PermissionConfigSchema, PermissionGroupSchema, PermissionSetupSchema, PermissionTypeSchema } from '../gameLogic';
-import { LIMIT_CHARACTER_PROFILE_LENGTH } from '../inputLimits';
+import { z } from 'zod';
 import { AssetPreferencesPublicSchema } from '../character/assetPreferences';
+import { CharacterPublicSettingsSchema, CharacterRoomPositionSchema } from '../character/characterData';
+import { CharacterIdSchema } from '../character/characterTypes';
+import { ChatCharacterStatusSchema, ClientChatMessagesSchema } from '../chat/chat';
+import { PermissionConfigChangeSchema, PermissionConfigSchema, PermissionGroupSchema, PermissionSetupSchema, PermissionTypeSchema } from '../gameLogic';
+import { AppearanceActionData, AppearanceActionProblem } from '../gameLogic/actionLogic/appearanceActionProblems';
+import { AppearanceActionSchema } from '../gameLogic/actionLogic/appearanceActions';
+import { LIMIT_CHARACTER_PROFILE_LENGTH } from '../inputLimits';
+import { Satisfies } from '../utility/misc';
+import { CharacterInputNameSchema, ZodCast } from '../validation';
+import { SocketInterfaceDefinition, SocketInterfaceDefinitionVerified, SocketInterfaceHandlerPromiseResult, SocketInterfaceHandlerResult, SocketInterfaceRequest, SocketInterfaceResponse } from './helpers';
 
 // Fix for pnpm resolution weirdness
-import type { } from '../assets/item/base';
 import type { } from '../assets/appearance';
+import type { } from '../assets/item/base';
 import type { } from '../character/pronouns';
 
 /** Client->Shard messages */
